requirements.txt
gradio
langchain
openai
python-dotenv
langchain_community
langchain_openai